Backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["hihumikan <a@qqey.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
mysql-connector-python = "^8.0.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
